+++ docs/_webCode/js/downloads.js
@@ -1,20 +1,20 @@
-// For this script to work on GitHub Pages, you need to ensure CORS is set up correctly on your Nexus server.
+// For this script to work on GitHub Pages, you need to ensure CORS is set up correctly on your Reposilite server.
 // For local testing, you need to disable CORS in your browser,
-// and you need to set the nexusUrl to the actual URL of your Nexus server including the target repository
+// and you need to set the repoUrl to the actual URL of your Reposilite server including the target repository
 // (like in your pom.xml).
-// E.g.: https://nexus.betonquest.org/repository/betonquest/
+// E.g.: https://repo.betonquest.org/betonquest/
 
 document$.subscribe(async () => {
-  const nexusUrl = "${REPOSITORY_URL}";
-  const parts = nexusUrl.split("/");
-  const baseUrl = parts.slice(0, -3).join("/") + "/";
+  const repoUrl = "https://repo.betonquest.org/betonquest/";
+  const parts = repoUrl.split("/");
+  const baseUrl = parts.slice(0, -2).join("/") + "/";
   const repositoryName = parts[parts.length - 2];
 
   window.onload = async function () {
     let urlParams = new URLSearchParams(window.location.search);
     const path = urlParams.get("path");
     if (path) {
-      const url = nexusUrl + path;
+      const url = repoUrl + path;
       const filename = urlParams.get("filename");
       await downloadWithRename(url, filename);
       window.location.href = window.location.href.split("?")[0];
@@ -24,44 +24,16 @@ document$.subscribe(async () => {
   await showBuilds();
 
   async function showBuilds() {
-    const builds = await getBuilds();
-
-    const cache = loadCache(builds);
-    await loadBuilds("release-build", builds.filter(build => !build.version.includes("-")));
-    await loadBuilds("development-build", builds.filter(build => build.version.includes("-")), cache);
-    saveCache(cache);
+    getBuilds("?snapshots=false").then(builds =>
+      loadBuilds("release-build", builds));
+    getBuilds("?releases=false&limit=100", true).then(builds =>
+      loadBuilds("development-build", builds));
   }
 
-  function loadCache(builds) {
-    let cache;
-    try {
-      cache = JSON.parse(localStorage.getItem("cachedVersions"));
-    } catch (e) {
-      console.log("Failed to parse cached versions");
-    }
-    if (!cache) {
-      cache = {};
-    }
-
-    if (builds) {
-      for (const version of Object.keys(cache)) {
-        if (!builds.some(build => build.version === version)) {
-          delete cache[version];
-        }
-      }
-    }
-
-    return cache;
-  }
-
-  function saveCache(cache) {
-    localStorage.setItem("cachedVersions", JSON.stringify(cache));
-  }
-
-  async function loadBuilds(idKey, builds, cache) {
+  async function loadBuilds(idKey, builds) {
     const latestBuild = document.getElementsByClassName("download-latest-" + idKey)[0];
     if (builds.length > 0) {
-      latestBuild.textContent = await getLazyVersion(builds[0], cache);
+      latestBuild.textContent = builds[0].version;
       const downloadUrl = builds[0].downloadUrl;
       latestBuild.onclick = function () {
         downloadWithRename(downloadUrl, "BetonQuest.jar");
@@ -74,14 +46,10 @@ document$.subscribe(async () => {
 
     const buildList = document.getElementById("download-all-" + idKey);
     resetDisabled(buildList.parentNode);
-    buildList.parentNode.addEventListener("click", () => {
-      const cache = loadCache();
-      loadBuildsLazy(builds, buildList, cache).then(() => saveCache(cache));
-    });
-
+    await loadAllBuilds(builds, buildList);
   }
 
-  async function loadBuildsLazy(builds, buildList, cache) {
+  async function loadAllBuilds(builds, buildList) {
     if (builds.length > 0) {
       const ul = document.createElement("ul");
       buildList.appendChild(ul);
@@ -89,7 +57,7 @@ document$.subscribe(async () => {
         const li = document.createElement("li");
         li.style.cssText = "padding: 0";
         const a = document.createElement("a");
-        a.textContent = await getLazyVersion(build, cache);
+        a.textContent = build.version;
         a.href = "#";
         a.onclick = function () {
           downloadWithRename(build.downloadUrl, "BetonQuest.jar");
@@ -103,83 +71,31 @@ document$.subscribe(async () => {
     }
   }
 
-  async function getLazyVersion(build, cache) {
-    if (!build.version.includes("-")) {
-      return build.version;
-    }
-    const nexusVersion = build.version;
-    if (cache[nexusVersion]) {
-      return cache[nexusVersion];
-    }
-    const result = await getPomResult(build);
-    if (!result) {
-      return null;
-    }
-    const version = result[1];
-    cache[nexusVersion] = version;
-    return version;
-  }
-
-  async function getPomResult(build) {
-    const pomUrl = build.downloadUrl.replace("-shaded.jar", ".pom");
-    const pomResponse = await fetch(pomUrl);
-    const pomData = await pomResponse.text();
-    return pomData.match(/<(?:plugin|betonquest)\.version>(.+?)<\/(?:plugin|betonquest)\.version>/);
-  }
-
   function resetDisabled(element) {
     element.style.pointerEvents = "auto";
     element.style.opacity = "1";
   }
 
-  async function getBuilds() {
-    let prereleaseSearch = undefined;
-    let continuationToken = undefined;
+  async function getBuilds(filter, firstGroupOnly = false) {
     const builds = [];
-
-    nextPage: while (continuationToken !== null) {
-      const params = getURLParams(prereleaseSearch, continuationToken);
-      try {
-        let data = await fetch(baseUrl + `service/rest/v1/search/assets?${params}`)
-          .then(response => response.json());
-
-        for (const build of data["items"]) {
-          const buildVersion = {version: build["maven2"]["version"], downloadUrl: build["downloadUrl"]};
-          if (buildVersion) {
-            builds.push(buildVersion);
-            if (prereleaseSearch === undefined && !buildVersion.version.includes("-")) {
-              prereleaseSearch = false;
-              continuationToken = undefined;
-              continue nextPage;
-            }
-          }
+    try {
+      let data = await fetch(baseUrl + `api/pommapper/id/BetonQuest` + filter)
+        .then(response => response.json());
+      for (const group of data) {
+        for (const versionEntry of group["versions"]) {
+          let pluginVersion = versionEntry["entries"]["pluginVersion"];
+          let betonquestVersion = versionEntry["entries"]["betonquestVersion"];
+          let downloadUrl = repoUrl + versionEntry["jar"].replace(".jar", "-shaded.jar");
+          builds.push({version: pluginVersion ? pluginVersion : betonquestVersion, downloadUrl: downloadUrl});
         }
-
-        continuationToken = data["continuationToken"];
-      } catch (error) {
-        console.error("Failed to fetch builds:", error);
-        continuationToken = null;
+        if (firstGroupOnly) {
+          break;
+        }
       }
+    } catch (error) {
+      console.error("Failed to fetch builds:", error);
     }
     return builds;
-  }
-
-  function getURLParams(prerelease, continuationToken) {
-    const params = new URLSearchParams();
-    params.set("repository", repositoryName);
-    params.set("group", "org.betonquest");
-    params.set("name", "betonquest");
-    params.set("maven.extension", "jar");
-    params.set("maven.classifier", "shaded");
-    params.set("sort", "version");
-
-    if (prerelease !== undefined) {
-      params.set("prerelease", prerelease);
-    }
-    if (continuationToken !== undefined) {
-      params.set("continuationToken", continuationToken);
-    }
-    return params;
   }
 
   function downloadWithRename(url, filename) {