+++ CHANGELOG.md
@@ -114,6 +114,7 @@ to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
     - `conversation.interceptor.display_history` config option, allow the interceptor to send the chat history after a conversation
 - `menu` `bound_item` now finds the first menu allowed to open instead aborting after the first check
 - `money` event now has a `multiply` argument instead of `*` in front of number
+- old maven artifact repository nexus `https://nexus.betonquest.org/repository/betonquest/` replaced with reposilite available at `https://repo.betonquest.org/betonquest/`
 ### Deprecated
 ### Removed
 - undocumented prefix feature in conversation